Fixes #2 Renamed directory app to src
Renamed app/ to src/ because this is not a app, it is a library and that makes more sense
