+++ package.json
@@ -31,7 +31,7 @@
   "scripts": {
     "test": "./node_modules/karma/bin/karma start",
     "live": "./node_modules/live-server/live-server.js",
-    "pro": "./node_modules/jspm/jspm.js bundle-sfx --minify app dist/bundle.js",
+    "pro": "./node_modules/jspm/jspm.js bundle-sfx --minify src dist/bundle.js",
     "clean": "./node_modules/gulp/bin/gulp.js clean",
     "compile": "./node_modules/gulp/bin/gulp.js compile",
     "compile-w": "./node_modules/gulp/bin/gulp.js compile-w",
